`vsl::forward`  parameters (#265)
* :truck: reorganize action modules for clarity.
	remove `vsl::app_dir` action since `toml::app.dirpath` is the same.

* :sparkles: add a forward target type to accept ip addresses when forwarding.
	   add a debug custom impl for Rcpt.

* :white_check_mark: add tests for ip6 with forwarding.

* :sparkles: add object support for the forward & forward_all actions.
	   added tests for those.

* :white_check_mark: add forward_all tests.

* :ambulance: fix lint issue.
