+++ vsmtp-rule-engine/src/tests/actions.rs
@@ -1,3 +1,5 @@
+use std::str::FromStr;
+
 /**
  * vSMTP mail transfer agent
  * Copyright (C) 2022 viridIT SAS
@@ -15,7 +17,9 @@
  *
 **/
 use crate::{rule_engine::RuleEngine, tests::helpers::get_default_state};
+use vsmtp_common::auth::Mechanism;
 use vsmtp_common::re::serde_json;
+use vsmtp_common::transfer::ForwardTarget;
 use vsmtp_common::{
     mail::Mail,
     mail_context::{Body, MessageMetadata},
@@ -31,7 +35,7 @@ fn test_logs() {
         &Some(root_example!["actions/logs.vsl"]),
     )
     .unwrap();
-    let mut state = get_default_state("./tmp/app");
+    let (mut state, _) = get_default_state("./tmp/app");
     assert_eq!(re.run_when(&mut state, &StateSMTP::Connect), Status::Deny);
 }
 
@@ -42,7 +46,7 @@ fn test_users() {
         &Some(root_example!["actions/users.vsl"]),
     )
     .unwrap();
-    let mut state = get_default_state("./tmp/app");
+    let (mut state, _) = get_default_state("./tmp/app");
 
     assert_eq!(
         re.run_when(&mut state, &StateSMTP::Delivery),
@@ -52,12 +56,8 @@ fn test_users() {
 
 #[test]
 fn test_send_mail() {
-    let re = RuleEngine::new(
-        &vsmtp_config::Config::default(),
-        &Some(root_example!["actions/send_mail.vsl"]),
-    )
-    .unwrap();
-    let mut state = get_default_state(format!("{}", root_example!["actions"].display()));
+    let (mut state, config) = get_default_state(format!("{}", root_example!["actions"].display()));
+    let re = RuleEngine::new(&config, &Some(root_example!["actions/send_mail.vsl"])).unwrap();
 
     // TODO: add test to send a valid email.
     assert_eq!(re.run_when(&mut state, &StateSMTP::Connect), Status::Accept);
@@ -70,7 +70,7 @@ fn test_context_write() {
         &Some(root_example!["actions/write.vsl"]),
     )
     .unwrap();
-    let mut state = get_default_state("./tmp/app");
+    let (mut state, _) = get_default_state("./tmp/app");
 
     state.get_context().write().unwrap().metadata = Some(MessageMetadata {
         message_id: "test_message_id".to_string(),
@@ -116,7 +116,7 @@ fn test_context_dump() {
         &Some(root_example!["actions/dump.vsl"]),
     )
     .unwrap();
-    let mut state = get_default_state("./tmp/app");
+    let (mut state, _) = get_default_state("./tmp/app");
 
     state.get_context().write().unwrap().metadata = Some(MessageMetadata {
         message_id: "test_message_id".to_string(),
@@ -153,7 +153,7 @@ fn test_quarantine() {
         &Some(root_example!["actions/quarantine.vsl"]),
     )
     .unwrap();
-    let mut state = get_default_state("./tmp/app");
+    let (mut state, _) = get_default_state("./tmp/app");
 
     state.get_context().write().unwrap().metadata = Some(MessageMetadata {
         message_id: "test_message_id".to_string(),
@@ -193,4 +193,202 @@ fn test_quarantine() {
         .expect("could not remove generated test file");
     std::fs::remove_file("./tmp/app/tests/generated/quarantine2/test_message_id.json")
         .expect("could not remove generated test file");
+}
+
+#[test]
+fn test_forward() {
+    let re = RuleEngine::new(
+        &vsmtp_config::Config::default(),
+        &Some(root_example!["actions/forward.vsl"]),
+    )
+    .unwrap();
+    let (mut state, _) = get_default_state("./tmp/app");
+    state.get_context().write().unwrap().body = Body::Parsed(Box::new(Mail::default()));
+
+    re.run_when(&mut state, &StateSMTP::Connect);
+    re.run_when(&mut state, &StateSMTP::Delivery);
+
+    let rcpt = state.get_context().read().unwrap().envelop.rcpt.clone();
+    println!("{rcpt:?}");
+
+    assert_eq!(rcpt[0].address.full(), "fqdn@example.com");
+    assert_eq!(
+        rcpt[0].transfer_method,
+        Transfer::Forward(ForwardTarget::Domain("localhost".to_string()))
+    );
+    assert_eq!(rcpt[1].address.full(), "ip4@example.com");
+    assert_eq!(
+        rcpt[1].transfer_method,
+        Transfer::Forward(ForwardTarget::Ip(std::net::IpAddr::V4(
+            std::net::Ipv4Addr::from_str("127.0.0.1").unwrap()
+        )))
+    );
+    assert_eq!(rcpt[2].address.full(), "ip6@example.com");
+    assert_eq!(
+        rcpt[2].transfer_method,
+        Transfer::Forward(ForwardTarget::Ip(std::net::IpAddr::V6(
+            std::net::Ipv6Addr::from_str("::1").unwrap()
+        )))
+    );
+    assert_eq!(rcpt[3].address.full(), "object.str@example.com");
+    assert_eq!(
+        rcpt[3].transfer_method,
+        Transfer::Forward(ForwardTarget::Domain("localhost".to_string()))
+    );
+    assert_eq!(rcpt[4].address.full(), "object.ip4@example.com");
+    assert_eq!(
+        rcpt[4].transfer_method,
+        Transfer::Forward(ForwardTarget::Ip(std::net::IpAddr::V4(
+            std::net::Ipv4Addr::from_str("127.0.0.1").unwrap()
+        )))
+    );
+    assert_eq!(rcpt[5].address.full(), "object.ip6@example.com");
+    assert_eq!(
+        rcpt[5].transfer_method,
+        Transfer::Forward(ForwardTarget::Ip(std::net::IpAddr::V6(
+            std::net::Ipv6Addr::from_str("::1").unwrap()
+        )))
+    );
+    assert_eq!(rcpt[6].address.full(), "object.fqdn@example.com");
+    assert_eq!(
+        rcpt[6].transfer_method,
+        Transfer::Forward(ForwardTarget::Domain("test.eu".to_string()))
+    );
+}
+
+#[test]
+#[allow(clippy::too_many_lines)]
+fn test_forward_all() {
+    let re = RuleEngine::new(
+        &vsmtp_config::Config::default(),
+        &Some(root_example!["actions/forward_all.vsl"]),
+    )
+    .unwrap();
+    let (mut state, _) = get_default_state("./tmp/app");
+    state.get_context().write().unwrap().body = Body::Parsed(Box::new(Mail::default()));
+
+    re.run_when(&mut state, &StateSMTP::Connect);
+
+    re.run_when(
+        &mut state,
+        &StateSMTP::Authentication(Mechanism::Login, None),
+    );
+
+    state
+        .get_context()
+        .read()
+        .unwrap()
+        .envelop
+        .rcpt
+        .iter()
+        .for_each(|rcpt| {
+            assert_eq!(
+                rcpt.transfer_method,
+                Transfer::Forward(ForwardTarget::Domain("localhost".to_string()))
+            );
+        });
+
+    re.run_when(&mut state, &StateSMTP::MailFrom);
+
+    state
+        .get_context()
+        .read()
+        .unwrap()
+        .envelop
+        .rcpt
+        .iter()
+        .for_each(|rcpt| {
+            assert_eq!(
+                rcpt.transfer_method,
+                Transfer::Forward(ForwardTarget::Ip(std::net::IpAddr::V4(
+                    std::net::Ipv4Addr::from_str("127.0.0.1").unwrap()
+                )))
+            );
+        });
+
+    re.run_when(&mut state, &StateSMTP::RcptTo);
+
+    state
+        .get_context()
+        .read()
+        .unwrap()
+        .envelop
+        .rcpt
+        .iter()
+        .for_each(|rcpt| {
+            assert_eq!(
+                rcpt.transfer_method,
+                Transfer::Forward(ForwardTarget::Ip(std::net::IpAddr::V6(
+                    std::net::Ipv6Addr::from_str("::1").unwrap()
+                )))
+            );
+        });
+
+    re.run_when(&mut state, &StateSMTP::Data);
+
+    state
+        .get_context()
+        .read()
+        .unwrap()
+        .envelop
+        .rcpt
+        .iter()
+        .for_each(|rcpt| {
+            assert_eq!(
+                rcpt.transfer_method,
+                Transfer::Forward(ForwardTarget::Domain("localhost".to_string()))
+            );
+        });
+
+    re.run_when(&mut state, &StateSMTP::PreQ);
+
+    state
+        .get_context()
+        .read()
+        .unwrap()
+        .envelop
+        .rcpt
+        .iter()
+        .for_each(|rcpt| {
+            assert_eq!(
+                rcpt.transfer_method,
+                Transfer::Forward(ForwardTarget::Ip(std::net::IpAddr::V4(
+                    std::net::Ipv4Addr::from_str("127.0.0.1").unwrap()
+                )))
+            );
+        });
+
+    re.run_when(&mut state, &StateSMTP::PostQ);
+
+    state
+        .get_context()
+        .read()
+        .unwrap()
+        .envelop
+        .rcpt
+        .iter()
+        .for_each(|rcpt| {
+            assert_eq!(
+                rcpt.transfer_method,
+                Transfer::Forward(ForwardTarget::Ip(std::net::IpAddr::V6(
+                    std::net::Ipv6Addr::from_str("::1").unwrap()
+                )))
+            );
+        });
+
+    re.run_when(&mut state, &StateSMTP::Delivery);
+
+    state
+        .get_context()
+        .read()
+        .unwrap()
+        .envelop
+        .rcpt
+        .iter()
+        .for_each(|rcpt| {
+            assert_eq!(
+                rcpt.transfer_method,
+                Transfer::Forward(ForwardTarget::Domain("test.eu".to_string()))
+            );
+        });
 }
